refine and update evaluate.py
