--- data/playwright/airbnb.txt
@@ -1,10 +1,0 @@
-page.goto("https://zh.airbnb.com/?_set_bev_on_new_domain=1701156118_MWJkOTM3Mzc2OTZh")
-page.get_by_role("button", name="下一个房源类目页面").click()
-page.get_by_role("button", name="下一个房源类目页面").click()
-page.locator("span").filter(has_text="魅力泳池").first.click()
-page.get_by_test_id("category-bar-filter-button").click()
-page.locator("#price_filter_min").click()
-page.locator("#price_filter_min").fill("200")
-page.locator("#price_filter_max").click()
-page.locator("#price_filter_max").fill("300")
-page.get_by_role("link", name="显示").click()
