--- data/playwright/google_play_card.txt
@@ -1,11 +1,0 @@
-page.goto("https://www.amazon.com/")
-page.get_by_placeholder("Search Amazon").click()
-page.get_by_placeholder("Search Amazon").fill("$100 Google Play Gift Card")
-page.get_by_role("button", name="Go", exact=True).click()
-page.get_by_role("link", name="Google Play gift code - give the gift of games, apps and more (Email or Text Message Delivery - US Only)").first.click()
-page.get_by_role("button", name="$100.00").click()
-page.get_by_placeholder("Enter an email for each recipient").click()
-page.get_by_placeholder("Enter an email for each recipient").fill("abc@abc.com")
-page.get_by_label("From").click()
-page.get_by_label("From").fill("Micheal")
-page.get_by_role("button", name="Add to cart").click()
